fix inconsistent behaviour after strike
On a strike, the module stops the sequence, and then resumes the
sequence after the strike animation, provided that the state hasn't
changed. It does this by accessing the DisplayedSequence field.

This field is not always indicative of the actual sequence(s) being
displayed. In particular, the Unstable quirk causes two sequences to be
played, only one of which is DisplayedSequence. The second sequence is
lost after a strike, even though the module is still in the Unstable
state.

This commit changes the strike mechanism to instead pause and resume the
displayed sequence, not interacting with the sequence itself in any way.