fix incorrect symbol sequence shifting
The action function corresponding to green action 4 ("Move all dots to the
start of the sequence, keeping the colors as they are.") assumes that the
number of dots is equal to the number of unique port types. This is only
necessarily true if this action was triggered directly by green
condition 4.

It is possible for this action to be triggered by other conditions
before condition 4, in which case this action function operates
incorrectly.

This commit changes this function to obtain the number of dots from the
sequence itself, removing the incorrect assumption.